requirements.txt
functions-framework==3.*
pandas==1.5.3
requests==2.26.0
google-cloud-storage==2.3.0
google-api-core==2.7.0
numpy==1.22.4